Fix uploading email attachments to sharepoint
This one was introduced in commit 301dd042fb0db917fe8e9a47f180722e6441af97.
Compare to the previous commit:

	git diff ffb88a1 301dd04

There was some adjustment of files and methods between these commits:

- in commit ffb88a1, in the file /app/case/views/case/email.py the
  email_attachment_view() function was the API endpoint used to handle email
  attachment download and delete. In both cases, these could only be done when
  the associated email was in the DRAFT state. The
  email_attachment_upload_view() function was the API endpoint used to handle
  email attachment upload to sharepoint. This did NOT require that the
  associated email was in a DRAFT state.

- in commit 301dd04, in the file /app/case/views/case_email.py the
  download_attachment_from_sharepoint() & delete_attachment() methods are used
  as the API endpoints used to handle email attachment download and delete
  respectively. The upload_attachment_to_sharepoint() method is the API
  endpoint used to handle email attachment upload to sharepoint.

Note that in the latter commit the guard condition to test if the associated
email was in the DRAFT state has gone from the delete_attachment() method and
has been added to a new method get_attachment() method. This method is used by
the upload_attachment_to_sharepoint() which did not previously (i.e. in the
previous commit email_attachment_upload_view() function) contain this condition.

The fix applied here was to move to position of the guard condition from the
get_attachment() method to the delete_attachment() method to replicate the
previous behaviour.

+++ app/case/views/case_email.py
@@ -241,9 +241,6 @@ class EmailApiViewset(GenericViewSet):
         return Response(serializer.data, status=201)
 
     def get_attachment(self, email: Email, attachment_id: int) -> EmailAttachment:
-        if not email.state == EmailState.DRAFT:
-            raise Http404()
-
         try:
             attachment = email.emailattachment_set.get(pk=attachment_id)
         except EmailAttachment.DoesNotExist:
@@ -259,6 +256,9 @@ class EmailApiViewset(GenericViewSet):
     )
     def delete_attachment(self, request, pk=None, email_id=None, attachment_id=None):
         email = self.get_email(email_id)
+        if not email.state == EmailState.DRAFT:
+            raise Http404()
+
         attachment = self.get_attachment(email, attachment_id)
         attachment.delete()
         return Response(status=204)
